requirements.txt
# LLM Document Processing System Dependencies

# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Gemini API
google-generativeai==0.3.2

# Neo4j Database
neo4j==5.15.0
py2neo==2021.2.4

# Environment and Configuration
python-dotenv==1.0.0

# HTTP and Networking
httpx==0.25.2
requests==2.31.0

# Data Processing and Utilities
numpy==1.25.2
pandas==2.1.4

# Text Processing and NLP
regex==2023.10.3
nltk==3.8.1

# Logging and Monitoring
structlog==23.2.0

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools (Optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1